requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
ojitos369==1.0
pytz==2024.1
setuptools==73.0.1
sqlparse==0.5.1
wheel==0.44.0
